requirements.txt
cryptography==41.0.3
PyJWT==2.8.0
argon2-cffi==21.3.0
bcrypt==4.